+++ client.go
@@ -4,8 +4,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"io"
+	"math/rand"
 	"time"
 
+	"github.com/oklog/ulid/v2"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/metric/instrument"
@@ -33,6 +36,8 @@ type Client struct {
 	backoff Backoff
 	meter   metric.Meter
 
+	entropy io.Reader
+
 	mEnqueue syncint64.Counter
 	mLockJob syncint64.Counter
 }
@@ -45,6 +50,9 @@ func NewClient(pool adapter.ConnPool, options ...ClientOption) (*Client, error) 
 		id:      RandomStringID(),
 		backoff: DefaultExponentialBackoff,
 		meter:   metric.NewNoopMeterProvider().Meter("noop"),
+		entropy: &ulid.LockedMonotonicReader{
+			MonotonicReader: ulid.Monotonic(rand.New(rand.NewSource(time.Now().UnixNano())), 0),
+		},
 	}
 
 	for _, option := range options {
@@ -87,17 +95,18 @@ func (c *Client) execEnqueue(ctx context.Context, j *Job, q adapter.Queryable) e
 		j.Args = []byte{}
 	}
 
-	err := q.QueryRow(ctx, `INSERT INTO gue_jobs
-(queue, priority, run_at, job_type, args, created_at, updated_at)
+	j.ID = ulid.MustNew(ulid.Timestamp(now), c.entropy)
+	_, err := q.Exec(ctx, `INSERT INTO gue_jobs
+(job_id, queue, priority, run_at, job_type, args, created_at, updated_at)
 VALUES
-($1, $2, $3, $4, $5, $6, $6) RETURNING job_id
-`, j.Queue, j.Priority, j.RunAt, j.Type, j.Args, now).Scan(&j.ID)
+($1, $2, $3, $4, $5, $6, $7, $7)
+`, j.ID.String(), j.Queue, j.Priority, j.RunAt, j.Type, j.Args, now)
 
 	c.logger.Debug(
 		"Tried to enqueue a job",
 		adapter.Err(err),
 		adapter.F("queue", j.Queue),
-		adapter.F("id", j.ID),
+		adapter.F("id", j.ID.String()),
 	)
 
 	c.mEnqueue.Add(ctx, 1, attrJobType.String(j.Type), attrSuccess.Bool(err == nil))
@@ -138,12 +147,12 @@ LIMIT 1 FOR UPDATE SKIP LOCKED`
 //
 // After the Job has been worked, you must call either Job.Done() or Job.Error() on it
 // in order to commit transaction to persist Job changes (remove or update it).
-func (c *Client) LockJobByID(ctx context.Context, id int64) (*Job, error) {
+func (c *Client) LockJobByID(ctx context.Context, id ulid.ULID) (*Job, error) {
 	sql := `SELECT job_id, queue, priority, run_at, job_type, args, error_count, last_error
 FROM gue_jobs
 WHERE job_id = $1 FOR UPDATE SKIP LOCKED`
 
-	return c.execLockJob(ctx, false, sql, id)
+	return c.execLockJob(ctx, false, sql, id.String())
 }
 
 // LockNextScheduledJob attempts to retrieve the earliest scheduled Job from the database in the specified queue.