+++ client_test.go
@@ -3,11 +3,14 @@ package gue
 import (
 	"context"
 	"errors"
+	"fmt"
 	"math"
+	"os"
 	"sync"
 	"testing"
 	"time"
 
+	"github.com/oklog/ulid/v2"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
@@ -35,7 +38,7 @@ func testLockJob(t *testing.T, connPool adapter.ConnPool) {
 	}
 	err = c.Enqueue(ctx, newJob)
 	require.NoError(t, err)
-	require.Greater(t, newJob.ID, int64(0))
+	require.NotEmpty(t, newJob.ID)
 
 	j, err := c.LockJob(ctx, "")
 	require.NoError(t, err)
@@ -48,7 +51,7 @@ func testLockJob(t *testing.T, connPool adapter.ConnPool) {
 	})
 
 	// check values of returned Job
-	assert.Equal(t, newJob.ID, j.ID)
+	assert.Equal(t, newJob.ID.String(), j.ID.String())
 	assert.Equal(t, defaultQueueName, j.Queue)
 	assert.Equal(t, JobPriorityDefault, j.Priority)
 	assert.False(t, j.RunAt.IsZero())
@@ -161,7 +164,7 @@ func testLockJobByID(t *testing.T, connPool adapter.ConnPool) {
 	}
 	err = c.Enqueue(ctx, newJob)
 	require.NoError(t, err)
-	require.Greater(t, newJob.ID, int64(0))
+	require.NotEmpty(t, newJob.ID)
 
 	j, err := c.LockJobByID(ctx, newJob.ID)
 	require.NoError(t, err)
@@ -174,7 +177,7 @@ func testLockJobByID(t *testing.T, connPool adapter.ConnPool) {
 	})
 
 	// check values of returned Job
-	assert.Equal(t, newJob.ID, j.ID)
+	assert.Equal(t, newJob.ID.String(), j.ID.String())
 	assert.Equal(t, defaultQueueName, j.Queue)
 	assert.Equal(t, JobPriorityDefault, j.Priority)
 	assert.False(t, j.RunAt.IsZero())
@@ -233,7 +236,7 @@ func testLockJobByIDNoJob(t *testing.T, connPool adapter.ConnPool) {
 	c, err := NewClient(connPool)
 	require.NoError(t, err)
 
-	j, err := c.LockJobByID(ctx, 0)
+	j, err := c.LockJobByID(ctx, ulid.Make())
 	require.Error(t, err)
 	require.Nil(t, j)
 }
@@ -258,7 +261,7 @@ func testLockNextScheduledJob(t *testing.T, connPool adapter.ConnPool) {
 	}
 	err = c.Enqueue(ctx, newJob)
 	require.NoError(t, err)
-	require.Greater(t, newJob.ID, int64(0))
+	require.NotEmpty(t, newJob.ID)
 
 	j, err := c.LockNextScheduledJob(ctx, "")
 	require.NoError(t, err)
@@ -271,7 +274,7 @@ func testLockNextScheduledJob(t *testing.T, connPool adapter.ConnPool) {
 	})
 
 	// check values of returned Job
-	assert.Equal(t, newJob.ID, j.ID)
+	assert.Equal(t, newJob.ID.String(), j.ID.String())
 	assert.Equal(t, defaultQueueName, j.Queue)
 	assert.Equal(t, JobPriorityDefault, j.Priority)
 	assert.False(t, j.RunAt.IsZero())
@@ -683,15 +686,12 @@ func testAdapterQuery(t *testing.T, connPool adapter.ConnPool) {
 
 	err = c.Enqueue(ctx, &j1)
 	require.NoError(t, err)
-	require.NotEmpty(t, j1.ID)
 
 	err = c.Enqueue(ctx, &j2)
 	require.NoError(t, err)
-	require.NotEmpty(t, j2.ID)
 
 	err = c.Enqueue(ctx, &j3)
 	require.NoError(t, err)
-	require.NotEmpty(t, j3.ID)
 
 	// test pool
 	testQueryableQuery(ctx, t, connPool, queue, &j1, &j2, &j3)
@@ -732,11 +732,11 @@ func testQueryableQuery(ctx context.Context, t *testing.T, q adapter.Queryable, 
 	err = rows.Err()
 	require.NoError(t, err)
 
-	assert.Equal(t, j1.ID, jobs[0].ID)
+	assert.Equal(t, j1.ID.String(), jobs[0].ID.String())
 	assert.Equal(t, j1.Type, jobs[0].Type)
-	assert.Equal(t, j2.ID, jobs[1].ID)
+	assert.Equal(t, j2.ID.String(), jobs[1].ID.String())
 	assert.Equal(t, j2.Type, jobs[1].Type)
-	assert.Equal(t, j3.ID, jobs[2].ID)
+	assert.Equal(t, j3.ID.String(), jobs[2].ID.String())
 	assert.Equal(t, j3.Type, jobs[2].Type)
 }
 
@@ -761,7 +761,7 @@ func TestMultiSchema(t *testing.T) {
 	}
 	err = c.Enqueue(ctx, newJob)
 	require.NoError(t, err)
-	require.Greater(t, newJob.ID, int64(0))
+	require.NotEmpty(t, newJob.ID)
 
 	j, err := c.LockJob(ctx, "")
 	require.NoError(t, err)
@@ -772,4 +772,65 @@ func TestMultiSchema(t *testing.T) {
 		err := j.Done(ctx)
 		assert.NoError(t, err)
 	})
+}
+
+func TestJobIDMigration(t *testing.T) {
+	connPool := adapterTesting.OpenTestPoolLibPQCustomSchemas(t, "job_id_migration_01", "job_id_migration_02")
+	ctx := context.Background()
+
+	// create a table with the serial ID
+	_, err := connPool.Exec(ctx, `
+DROP TABLE IF EXISTS gue_jobs;
+CREATE TABLE gue_jobs
+(
+  job_id      BIGSERIAL   NOT NULL PRIMARY KEY,
+  priority    SMALLINT    NOT NULL,
+  run_at      TIMESTAMPTZ NOT NULL,
+  job_type    TEXT        NOT NULL,
+  args        BYTEA       NOT NULL,
+  error_count INTEGER     NOT NULL DEFAULT 0,
+  last_error  TEXT,
+  queue       TEXT        NOT NULL,
+  created_at  TIMESTAMPTZ NOT NULL,
+  updated_at  TIMESTAMPTZ NOT NULL
+);
+
+CREATE INDEX IF NOT EXISTS idx_gue_jobs_selector ON gue_jobs (queue, run_at, priority);
+`)
+	require.NoError(t, err)
+
+	// insert several records to test if the data migration works fine
+	now := time.Now()
+	const queueName string = "some-queue"
+	for i := 0; i < 101; i++ {
+		_, err = connPool.Exec(ctx, `INSERT INTO gue_jobs
+(queue, priority, run_at, job_type, args, created_at, updated_at)
+VALUES
+($1, $2, $3, $4, $5, $6, $6)
+`, queueName, 0, now, "foo-bar", []byte(fmt.Sprintf(`{"job":%d}`, i)), now)
+		require.NoError(t, err)
+	}
+
+	migrationSQL, err := os.ReadFile("./migrations/job_id_to_ulid.sql")
+	require.NoError(t, err)
+	_, err = connPool.Exec(ctx, string(migrationSQL))
+	require.NoError(t, err)
+
+	// ensure it is possible to retrieve a job from the DB after the conversion
+	c, err := NewClient(connPool)
+	require.NoError(t, err)
+
+	j1, err := c.LockJob(ctx, queueName)
+	require.NoError(t, err)
+	require.NotNil(t, j1)
+	t.Logf("Locked a job: %s %s", j1.ID.String(), string(j1.Args))
+
+	err = j1.Delete(ctx)
+	require.NoError(t, err)
+	err = j1.Done(ctx)
+	require.NoError(t, err)
+
+	j2, err := c.LockJobByID(ctx, j1.ID)
+	require.Error(t, err)
+	require.Nil(t, j2)
 }